Add files via upload
A image detection program that i followed a youtube tutorial for cv2 guide and changed some of the coding to fits the game i want. For this instance, is Arcane Odyssey for roblox which allow you to afk and auto fish. (Found out is not allowed in game so didn't get to use it ._.)